Rename coreos module to aleph
Pure rename of coreos.rs -> aleph.rs and update of all import
paths. No functional changes, done as a separate commit so git
tracks the rename and subsequent changes show as a diff.

Authored-by: Opencode.ai claude-opus-4-6@default
